Created dashboard module which manages bokeh plots
Created Dashboard class which create dashboard showing table of input
data and plot of all transactions.
Updated inputs to calculate absolute amount of transaction and colour
column for positive/negative transactions.
Updated main to create the Dashboard and allow the program to be ran
with `bokeh serve --show src\banalysis`.